Resolve lint warning
