+++ packages/profile/src/components/provider/theme.tsx
@@ -1,5 +1,5 @@
-import { createContext, useContext, useEffect, useState } from "react";
-import { useQueryParams } from "./query";
+import { createContext, useEffect, useState } from "react";
+import { useQueryParams } from "./hooks";
 
 type Theme = "dark" | "light" | "system";
 
@@ -19,7 +19,8 @@ const initialState: ThemeProviderState = {
   setTheme: () => null,
 };
 
-const ThemeProviderContext = createContext<ThemeProviderState>(initialState);
+export const ThemeProviderContext =
+  createContext<ThemeProviderState>(initialState);
 
 export function ThemeProvider({
   children,
@@ -70,12 +71,3 @@ export function ThemeProvider({
     </ThemeProviderContext.Provider>
   );
 }
-
-export const useTheme = () => {
-  const context = useContext(ThemeProviderContext);
-
-  if (context === undefined)
-    throw new Error("useTheme must be used within a ThemeProvider");
-
-  return context;
-};
